requirements.txt
APScheduler==3.10.4
Flask==3.1.0
python-dotenv==1.0.1
gspread==6.1.4
oauth2client==4.1.3
reportlab==4.2.5
requests==2.32.3
pytz==2024.2
